requirements.txt
click
requests
numpy
Pillow
scipy
pyspng
torch==1.13.1
torchvision==0.14.1
tqdm
